requirements.txt
Flask
Flask-Cors
pandas
openai
python-dotenv
